Crate fedimintd

Modules

envs

metrics 🔒

Structs

ServerOpts 🔒

Constants

SHUTDOWN_TIMEOUT 🔒

Time we will wait before forcefully shutting down tasks

Functions

default_esplora_server

default_finality_delay 🔒

For real Bitcoin we want to have a responsible default, while for demos nobody wants to wait multiple minutes

default_modules

run

Block the thread and run a Fedimintd server