Module fedimintd::encrypt

Encrypt and authenticate data stored on the filesystem with a user-supplied password.

We encrypt the configs to prevent attackers from learning the private keys if they gain file access. We authenticate the configs to prevent attackers from manipulating the encrypted files.

Users can safely back-up config and salt files on other media the attacker accesses if they do not learn the password and the password has enough entropy to prevent brute-forcing (e.g. 6 random words). Switching to a memory-hard algorithm like Argon2 would be more future-proof and safe for weaker passwords.

We use the ChaCha20 stream cipher with Poly1305 message authentication standardized in IETF RFC 8439. PBKDF2 with 1M iterations is used for key stretching along with a 128-bit salt that is randomly generated to discourage rainbow attacks. HMAC-SHA256 is used for the authentication code. All crypto is from the widely-used ring crate we also use for TLS.

Constants

CONFIG_FILE

DB_FILE

ITERATIONS 🔒

SALT_FILE

TLS_CERT

TLS_PK

Functions

encrypted_read

Reads encrypted data from a file, returns an incremented nonce for encrypting the next file

encrypted_write

Write data encrypted to a file with an unused nonce that will be encoded in the file

get_key

increment_nonce 🔒

returns a nonce from bytes and an incremented nonce for encrpyting the next message

zero_nonce