fedimintd

Function run

pub async fn run(
    module_init_registry: ServerModuleInitRegistry,
    code_version_hash: &str,
    code_version_vendor_suffix: Option<&str>,
) -> !

Block the thread and run a Fedimintd server

Arguments

modules_fn - A function to initialize the modules.

code_version_hash - The git hash of the code that the fedimintd binary is being built from. This is used mostly for information purposes (fedimintd version-hash). See fedimint-build crate for easy way to obtain it.

code_version_vendor_suffix - An optional suffix that will be appended to the internal fedimint release version, to distinguish binaries built by different vendors, usually with a different set of modules. Currently DKG will enforce that the combined code_version is the same between all peers.